Use PBF as main OSM format and improve README.md
When --analyze is used, look for a PBF file with the OSM data,
if this is missing look for .osm, if this is missing too
tell the user to read README.md.
Improve README.md so that the user know that a manually
created OSM file must be saved as `project_dir/data/osm_data/task/task.pbf`.
Fix typo in README.
Requested/reported by sabas in #1.